Supported margin style props
Have to apply them manually to the LayoutParams otherwise they have no effect. Also have to requestLayout of the CoordinatorLayout so FAB is redrawn in new position (same as when changing gravity).
Applied margins after all props updated otherwise didn't work when changing from margin: 10 to marginLeft: 10, for example - the change to marginLeft happened and then change to margin (defaults to 0) happened so marginLeft got overwritten