Supported RTL languages in FAB
Added start/end (as well as left/right) to gravity and margin props. For example, if RTL language a bottomEnd gravity will be on the left (but a bottomRight will always be on the right).
Had to set the layoutDirection on the CoordinatorLayout - tried it on the FloatingActionButton and it didn't work?!