Added icon and tint props to BottomAppBar
Copied them over from the Toolbar component.
Got stuck for a while because the icons didn't have the correct height (should be 56dp). Found out that the mMaxButtonHeight of the toolbar wasn't initialised correctly - it was -1 but should be 154 (56dp converted to pixels). It was set correctly for the toolbar but not the bottomappbar which extends toolbar!?!
The single arg BottomAppBar constructor passes -1 as the defaultStyleAttr so it wasn't finding the toolbar style - change to callt he double arg constructor which correctly passes R.attr.bottomAppBarStyle as the defaultStyleAttr