Rendered bottom app bar via navigation bar
This was the original idea but got put off cos bottom app bar doesn't have any navigation items. But they're the same component at heart so makes sense to combine.
Had to render bottom app bar before search because so that the ref points at app bar (react native uses the first component as the ref). Probably creating trouble because fabric doesn't work like that but worry about that when the time comes