Deleted hide prop from floating action button
No good having a prop when it can change on native side only. For example, could set hide to true then it could be shown on client by scrolling down the navigation bar - but can't hide it again because hide prop already true. Would either have to add imperative hide instance method or make it a controlled prop and update it on react side