Added anchor prop to extended fab
What should happen is, the fab should shrink as the component it's anchored to moves. But the shrink animation wouldn't work. It would either jump around or disappear and come back, depending on where the anchor gravity. Debugged by creating a new java project using the Scrolling Activity template and compared to the twitter sample with collapsing navigation bar. Eventually found that the measure/layout of the fab is called in the template project but not the twitter one. When the shrink animation starts then measure/layout is called by the java choreographer - but it isn't called in the twitter one. Have no idea why, but know that react native has its own choreographer?! Anyway, found that by manually calling the measure/layout in request layout fixes the shrink animation. Not meant to call measure/layout from requestLayout but only calling it on the fab so it isn't expensive - calling it higher up the tree would be expensive.