Ensured fab shrunk when anchored to nav bar
Checked against java sample and when anchored to a toolbar (not collapsing0 then the extended fab is always shrunk. The logic that calculates this is the getMinimumHeightForVisibleOverlappingContent method of the appbarlayout. This method checks the height of the last child.
So can't have the last child being the status bar because this is 0 height and meant that the extended fab would start off extended