Moved status bar out of nav bar on android
Want the calculation in getMinimumHeightForVisibleOverlappingContent of appbarlayout to be correct so have to have the toolbar as the last child of the appbarlayout - this is what decides when to start shrinking the extended fab when nav bar collapses.
Hard-coded the width and height of status bar to 0. It will be 0 if outside of the coordinatorlayout because that comes from react native. But inside the coordinatorlayout the width and height come from android so forced it to be 0.
Kept it inside the navigation bar on ios because that already works and ios searches for it by tag name so narrow the search by starting at the navigation bar for speed