Primed context for fluent peekable on ios
If fluently navigate from A to A -> B -> C -> D then need to render C on iOS (even though on D). Otherwise the view isn't loaded in time (in viewLoad) when peeking back - this causes problems, particularly for the header where the title and search bar don't load, but also shows a flash of a blank scene.
So load the previous scene C when on D so that it's loaded when peeking or navigating back. Will need native changes to block back navigation until loaded (should be quick so user would have to press back v v fast twice to be blocked)