Blocked back until scene rendered
When fluently navigating from A to A -> B -> C  then scene B isn't loaded right away. If allow back navigation to it then the navigation bar doesn't work (missing title, no search bar) because ios expects view to be loaded in loadView of UIViewController. So block back navigation until the scene is rendered - this happens in a setTimeout to not impact the rendering of scene C. Blocked it using a fluent prop on scene - only allow back if this is false.
Tried to run set a peekable field on NVStackController from checkPeekability and return this from shouldPopItem. But if pressing back fast then the shouldPopItem is called before anything else (like didShowViewController). So played it safe and worked out peekability from the navigationItem passed to shouldPopItem.
To really test it need to fluently navigate from A to A -> B -> C -> D and then press back twice fast