Revalidated fluent peekable in timeout
Upped timeout to 10 secs for testing. Say fluently on A -> B -> C so that B is fluently peekable but after 5 secs navigate to A -> B -> C -> D so that B isn't peekable anymore. Wait for 7 secs and then go back. Shouldn't have rendered B because no point doing a set state until it's peekable.
This does reset the timer so have to wait another 10 secs to go back to B. But in reality the timeout isn't 10 secs it's 0 so this is ok