Supported search bar on fluent navigation
Thought it would work but it didn't?! Tried fluently navigation from A to A -> B -> C -> D with search bar on B and C. Then pressing back fast showed search bar on C but not B!!!
Added logging and it seems that viewDidLoad fired too soon, even before the scene is navigated to. B's didLoad fired when on C. Even though couldn't navigate back to it, the viewDidLoad fired and B's subviews weren't loaded yet. So there wasn't any search bar to find in the viewDidLoad and it didn't get loaded.
Moved the logic from viewDidLoad to viewWillAppear and it works. This only fires if it will actually appear - after the back button check that it's rendered so subviews are guaranteed.
Simplified the check - used subviews.count instead of fluent prop. It's clearer and anyway the props are set before subviews loaded so not safe to use prop