Prevented freeze if peek navigation
If fluently navigate from A to A --> B --> C then B hasn't rendered anything. So can't freeze because B will be blank if peeking (this actually prevents peeking because back is blocked until B is rendered). This means B will be unfrozen for good so will render even though on C - not optimal