Navigated peek whenever native nav first
If pressing/swiping back or double tapping tab then the native navigation happens first. Must unfreeze so the user sees the latest render with all the state updates (e.g. redux). Simplified by always creating a peek navigation event when native navigation first - used the history flag to determine it's a native first navigation. And never freeze if peek in state context.
If navigate from A to A --> B and then peek back then A will be unfrozen - even if don't complete the peek/swipe back. So it's not optimal but that's ok