Froze peek if not previous crumb (or first)
Navigate from A to A --> B --> C then peek back to B. So B will be unfrozen. But then don't complete the swipe and navigate to D instead. So B is still unfrozen. If peek back to C then B and C are unfrozen together.
To avoid this only unfroze the peek if it's the previous scene. This means that B will refreeze when navigate to D.
Had to also unfreeze if the first scene because double tap tab peeks the first scene. Want this to unfreeze so the first scene updates immediately when double tap a tab