Created a peek unfreeze window
There was still a problem with previous approach. Navigate from A --> B --> C and peek back to B but don't complete swipe. Then B unfreezes. Then navigate to D and then B will freeze again, but when navigate back to C B would unfreeze again because it is the previous crumb. Also the problem that had to keep the first crumb unfrozen after peeking back to allow for unfreezing when double tapping a tab.
So instead stamped the peeked context with a timestamp and then only unfroze if within that time period. The time period check only runs when there's a navigation. So if on A --> B --> C and peek back to B then B is unfrozen and stays unfrozen while on C. But as soon as navigate to D then the timestamp is checked and B freezes again. Even if then navigate back to C then B stays frozen