Ensured freeze only changes after navigation
Using timestamp wasn't completely reliable. For example, could be on A --> B --> C and peek back to B and very fast navigate to D. Then B would still be unfrozen because the navigation happens before the freeze window elapse. Also, possible that the whole app re-renders outside of a navigation - maybe only in theory - then the window might elapse but still might be peeking. So should stay unfrozen but would freeze then.
By stamping the peek with the navigationEvent from props then will freeze again whenever a navigation happens - just right. So A --> B --> C and peek back to B. Then B will be unfrozen and stay unfrozen until a navigation happens. But will freeze and stay frozen after any navigation. So navigate to D and even back to C but B will stay frozen. As soon as navigate to B then it will obviously unfreeze because it will be the current crumb