Prevented freezing until animation completes
Was freezing immediately on the javascript side, as soon as the navigation happened. This worked but think only because of React Native bug with display none not working if position absolute, https://github.com/facebook/react-native/issues/18415. When suspended React Native sets display none on the scene, so it should be hidden during the animation - but it wasn't because of this bug. Can see there's a Yoga fix for it so can't leave it like this. Instead have to wait for the animation to complete before freezing. Wanted to avoid this because it's harder but it's the right thing to do. Also gives the unloading and unloaded time to do a setState because should fire fast enough, before animation completes.
Navigate A --> B --> C so that A and B are frozen. When navigating back to B the freeze is disabled during the navigation - but interestingly A stays frozen! Was surprised but think that's because the shouldComponentUpdate returns false - the old way of preventing a setState!! so it correctly doesn't unfreeze!! Nice that the old setState prevention comes into play for the new freeze way during the animation!!
Only fired the onRest for the entering scene. Even in odd case where leaving animation is longer than entering, as soon as entering completes it's ok to hide the leaving because can't see it anymore.
Called it onRest to match the NavigationMotion - it also tracks the animation complete, also only on the mounted/entered scene.