Froze when no animation
If unmountStyle={() => ''}, so there's no animation for the new scene, still need to call onRest. Otherwise the previous scene isn't frozen. Called onRest straight away - checked if the fragment is resumed for safety. Mustn't call onRest if createAnimation called when fragment paused for example. Better not to call onRest than to call it when shouldn't - can under freeze but mustn't overfreeze