Prevented freeze until animation completes (ios)
Same change as made for android but don't need to listen to animation - used didShowViewController instead. Was already using it for tell javascript about native navigation (back/swipe/tab) so repurposed as onRest and used for both. Only pass an eventCount if it's a native navigation that the javascript has to catch up with. Don't need an eventCount otherwise because only notifying that the scene animation has completed.
Removed hidden override from scene now waiting for animation to complete. Don't trust the setHidden override - think it only works because there's a React Native but with display none for position absolute. So don't want to rely on it