Prevented React Native hiding frozen scenes
Navigate A --> B --> C then A and B are both frozen. Then press back on iOS twice quickly and screen A will be black because the scene is hidden. When scenes are frozen/suspended then React Native hides them by setting display to none. Pressing back fast to A will unsuspend A in javascript but not immediately - the navigation has to happen in javascript with a setState etc. In the meantime A is shown and it's still hidden!!
Tried so many ways around this and almost ran out of options. One thing that did work was to override setHidden on SceneView and ignore true value. But this is taking advantage of a Yoga bug where display none doesn't work properly with position absolute https://github.com/facebook/react-native/issues/18415 - can see that it's been fixed in Yoga and then overriding setHidden won't work because the frame of the children is set to 0 (I think?!). Tried it without position absolute and it didn't work.
So have to stop React Native setting display to none, https://github.com/facebook/react-native/blob/1465c8f3874cdee8c325ab4a4916fda0b3e43bdb/Libraries/Renderer/implementations/ReactNativeRenderer-prod.js#L5480. Only way I found is to make display an invalid style attribute. That way the display none is ignored!!! Got lucky with but think it's robust enough to merge. Think when fabric lands the valid attributes will get overhaul - might make it easier to define own list of valid attributes?! but might break this workaround. Who knows?!
Tested this doesn't break display attribute for other non-Scene components