Prevented freeze of exiting scene
A --> B --> C then ios swipe back to C. This didn't fire useEffect after useUnloaded. That's because the native navigation happens first so don't set rest to false like when navigating in javascript first. So the freeze happens straight away, at same time as javascript navigation so there's no time for a useEffect if setState done in useUnloaded. Want to allow this - not just because it used to work but also there's no need to freeze the leaving scene because it's not in the stack anymore anyway