Prevented over-rendering when no suspense
Tested by setting React.Suspense to undefined so it's like react < 17. If freezable is turned on then the crumb scenes render on every navigation so A --> B --> C renders A, B and C. But should only render C. That's how it used to work. So turned off freezable if there's no suspense so the shouldComponentUpdate behaves as before.
If Suspense is available it's ok to re-render A and B because they're frozen.