Matched up custom bar buttons on ios and android
Can't make them behave exactly the same because on ios the bar button is sized to match the width of the contents but on android need to set an explicit width (wrap contents doesn't work on react native - maybe there's another way, but safest to set a width).
So added the bar button view as the custom view on ios - same as android - instead of first child view. This means can have multiple children and matches title view. Can't break existing custom view on ios so can't set a width.
Also centered the content on android so that the default behaviour on android works. Also makes ios and android work the same if user sets a width and a height on ios too