Fixed dynamic custom view buttons on ios
Started with a custom view button and then added one after a while and the first custom view button disappeared (it was there but not rendered and reappeared after navigating forward and back). This came in when made the custom view the self of the NVButtonView - doesn't happen if it was the first subview. Want to keep it as self so that it matches with android.
So prevented the NVButtonViews being added to the parent NVBarView. Not sure why this caused problem but ensures they only don't have a parent when set to the custom view