Gave android button sensible defaults
The button bar view is only part the custom view container on android - couldn't make this work on ios because it broke ref for alertios. So gave container good defaults for android - need a width for it to work, and centered so that an image appears where it should (matches built in menu items).
Allowed the width to be overridden (size prop), but only the width. Can always style as required by adding a child view and putting styles on that. That makes android and ios equivalent