Changed to match native behaviour
Was too confusing trying to incorporate navigation. Changed so that it behaves the same as the native tabbar. Then can progressively enhance with hyperlinks like with other navigation. For example, add an href to tabbaritem. Then in userland can turn the tabbar into a controlled component and set the tab prop from the navigation data