Rendered using AppRegistry as per rnw docs
Want it to just work as per the react native web docs. So don't want anything in index.web.js except the AppRegistry stuff from react native web. So exported NavigationMotion as NavigationStack and gave it default slide animation and it works! Had to prevent the notificationsNavigator navigate otherwise it ran last and always started on notifications state.