Allowed current scene to re-render
If update renderScene prop on NavigationStack then need to re-render the current scene (typical scenario is passing props from parent scene to child Modal stack - see issue #554). But the shouldComponentUpdate was stopping this from happening. On Android the freezable prop was set to false for the first scene - the first scene has a lifecycle of STARTED but was checking for RESUMED and so rest wasn't update to true. Changing to STARTED does update the scene, but want this to work without suspense too (non-freezable) - not (just) because want to support React before Suspense, but also because don't want to rely on Freeze. Freeze relies on React Native internals so don't want to rely on it. So allowed the render if it's at rest and it's the current scene or suspense available