Went back in history until url is reached
Take the scenario where navigate from A to A --> B. There's 1 back item in browser history for A. Then navigate within B (without adding to crumb trail). There's now 2 back items in history - one for B and one for A. So when navigateBack(1) actually need to go back twice in history!
The problem is history.back isn't synchronous so can't just keep calling history.back() and checking reached the right url. Instead have to go back the expected distance and then check if it'st he right url in the popstate listener