Prevented wrong scene showing during unmount
When navigating from A → B then back to A, then could get scene A showing in place of scene B. So scene A would show twice - both as the mounted and the unmounting scene. Scene B calculated its state and data from the global stateNavigator in props instead of the local stateNavigator in state. It only showed up when the mounted and crumb styles were the same so that onRest fired before B unmounted and triggered B to rerender.