Prevented rendering scenes while animating
Navigate from A --> B back to A. If the mounted and crumb styles are the same then rest is set to true immediately. The mounted scene A is immediately at rest because there's nothing to interpolate. This caused scenes to render while ongoing B animation plays - because the shouldComponentUpdate of Scene returns true when rest is true.
Added a marker to style to ensure there's always something to interpolate. Ensures that rest isn't set before duration elapses. The marker is 1 for unmounted, 0 for mounted and -1 for crumb