Added defensive check for fluent skip
If fluently navigate from A to A --> B --> C then navigate back 2 then stateContext is null and crumbs aren't long enough to provide state and data on B. Found this bug on Mobile and tried and failed to recreate on native. Although can't recreate it still adding the defensive check. It keeps it in step with the mobile code and there are probably some edge cases where it could happen?!