Stayed on same scene when refreshing!
When doing mobile or native web navigation, often want to navigate and stay on the same scene. For example, changing colors on zoom sample or expanding thread on email sample. Always had to truncate crumb trail (or fluently navigate back then forward!?). This is complicated.
Changed refresh navigation to stay on the same scene! That's what refresh navigation is meant to do anyway. It's odd that it pushed a new scene on the stack when it's called refresh. If want to push then call navigate.