Smoothed animation when new scene renders slowly
When navigating to a new scene that takes a while to render the animation would stutter. Tested with a 1 second sleep in the render and there is no animation at all - the new scene appears immediately after the sleep.
It's because the first animation tick is set straight away and the second tick doesn't come in until after the render completes. So if the render takes half the animation duration then it will jump.
Fixed by capping the progress time between ticks. So even if the ticks are 300ms apart, capped the delta at 50ms. So the animation runs smoothly after the animation completes. In fact, it behaves more like native where there's a slight delay at the beginning of the transition while the scene renders and then the animation runs smoothly