Prevented flash when opening modal stack
The NavigationStack background color was hard-coded to black. It was visible briefly when opening a stack in a Modal. Added an underlayColor prop to the NavigationStack so it can be set to a transparent color, e.g., rgba(0,0,0,0). Then can set the Modal to transparent and this removes the flash. Also allows the background scene to show through in medley example when navigating - instead of a black background which made it look like it wasn't a modal