Prevented rendering current when slow scene leaves
Navigate to A --> B --> C and then navigate back to A --> B. If C transition duration is 3s but B duration is 1s then B will keep rendering after 1s - for the 2s extra it takes C to complete, B will keep rendering. That's because B is at rest and is current so will render.
So changed to only set to rest when all scenes are at rest. Can also do away with rest in state because can calculate if the mounted scene is at rest in similar way