Cleared context of invalid urls
The old and previous urls won't be valid when updating the state navigator with new states.
Added defensive check to scene because state will be null when context is cleared.
Would need to properly resolve these issues if doing component api for real. For example, could navigate and update children in between so it will navigate to an invalid url when animation completes