Kept old states around but marked as deleted
If they're not around then things get weird. There are still old scenes with old state context that try to parse old links. So best if these old links still parse.
So added the old states to the end of the array (give precedence to new states so they match routes first).
But kept getting weird errors if change the stack at same time as navigate in the UI. The stack changes then the navigation comes in while the stack is animating the replace. It processes the new navigation but that's against the old state context - so the old crumbs are wrong. Think that's what's happening. So validated that the state context isn't stale! Checked if the state on the state context is still in the latest states!