Prevented blank scenes when navigation backs up
Say you're on A --> B and you click a button to go to scene C. But before the C navigation takes place there's a navigation from React side that goes back to A.
The navigation back to A happens first on the React side followed by the navigation forward to C.
Then the navigation back to A happens on the native side followed by the navigation forward to C.
But when the navigation back to A completes the scene B is popped. This pop sends an event back to React to remove scene B. Then when the navigation forward to C completes scene B is still in the stack on the native side but it has been removed on the React side. So if you press back now you'll see a blank scene B!!!
This is because scenes didn't have a unique key - they had the key of their position in the stack. So scene A has key 0, B has key 1, C key 2 etc. Scene B always has the same key, so when the first B is popped on the native side all scene B's are removed on the React side.
Changed to give scenes a unique key by prefixing with a counter. So A has key 0-0 then B has 1-1. Then when navigating back to A and then forward to C then the stack will be 0-0 --> 3-1 --> 3-2. Going back to A increments the counter to 2. Scene A keeps its key of 0-0. Can't change the key of scenes already in the stack otherwise React would unmount them. Then when the forward to C happens both B and C get new keys with a counter of 3.
This time when the animation back to A completes the first B is popped and React removes the scene with key 1-1. So you can still go back from C to B because B with key 3-1 is still there in React!