Prevented error when navigation backs up
Repeated same test on ios and can get error when the navigation backs up. The test is a navigation back to home that runs every 4s (with a javascript sleep to create backup) and then navigated from A --> B to A--> B --> C but before it happens it navigates to home. Then when the navigate to C happens and then press back button it threw an error at objectAtIndex call because keys weren't big enough.
Not sure why?!