Implemented enter anim prop
Can't have an empty string in the new arch - both undefined and empty string come across as empty sting. So had to add a enterAnimOff bool that indicates if it's empty.
Tool me ages to work out that the prop has to be specified in the Manager.m obj-c class otherwise it isn't set?! So the managers are still relevant in the new architecture