Support reload ('r' in console)
Not sure if this is right but seems to work.  Had to remove scene from superview when it's unmounted otherwise couldn't recycle it. Also cleared view controllers so it starts over - tried other things like popping to root but they weren't reliable. Reset native event count because mostResetEventCount resets to 0. Lastly added defensive check in willShowViewController because the first navigation is for a generic UIView?!