Accepted title font props
Added default value of -1 for font size because can't get a value of nil when props are undefined?! Then set to nil on the native side.
Decided not to check for props change and just set the properties anyway. Otherwise native properties start as nil but the incoming props don't