Versioned progress as wip so can report bug
Gonna report a bug to the React Native Working Group about contentInsetAdjustmentBehavior failing. If navigate from A to B then the scrollview on B shows correctly below the navigationbar. But if navigate back then forward to B again then the scrollview starts at top of screen, under the navigationbar?! Don't know why but suspect it's because of the view recycling that seems to be happening in fabric