Preferred not to use oldViewProps and just update
Not sure about using oldViewProps with view recycling in the new arch. For example, could be a recycled view with font matching the font of the last time the view was used - so the props match! but still need to update the button's font because the button is recycled and created new