Defined scenes statically and render dynamically
Finally worked out the component api!! The key is to define the StateNavigator (scenes) statically and only the renderScenes dynamically. Need to keep scenes static because otherwise mutate the StateNavigator. Mutating makes the code complicated and doesn't work for concurrent rendering (React 18). The mutation has to happen before the render so can't render in the background.
Also only want the render to be inside the component. That's the component api. Static scenes and dynamic render is win/win