Ensured work with old navigation react mobile
Should still work with old navigation react mobile. Can't move children to renderMotion because that's a new prop. So left children were they are. Also only set renderMotion if children is not a function. Can't set it always because renderMotion takes precedence over children so the default renderMotion would run instead of children