Removed redundant initial navigate
The component api handles it now, navigating to the first scene if the state context is blank.
Had to set the stateNavigator in the new state otherwise ended up in an infinite set state loop - the check to stop an infinite loop never went to true. This happens now moved getDerivedStateFromProps to setState with hooks