Prevented error on remount nested NavigationStack
The test is have a scene with a NavigationStack - so a nested NavigationStack. Then put a useEffect in this scene - it doesn't have to contain anything, just an empty one with empty dependency array. Then delete the useEffect and get a black screen!!
Not sure what's going wrong but something doesn't like an empty view controller array. So put a dummy element in and then removed it later.