Prevented UIViewControllerHierarchyInconsistency
Sometimes got error: 'UIViewControllerHierarchyInconsistency', reason: 'A view can only be associated with at most one view controller at a time!. Best guess is that React Native is trying to reuse a View that'. Only when doing the test in the previous commit where toggling useEffect in a nested NavigationStack.
Best guess is that React Native is reusing the scene before it's remove the the view controller?!