Survived live/reloads with nested NavigationStack
Been investigating this for about 5 days straight. The setup is nested NavigationStack and a useEffect in the nested stack. Then removing the useEffect and saving, and adding it back and saving. Just couldn't get it working - would show a black screen. If the nested stack was in a TabBar then would get a white screen. Sometimes reloading works (typing R), sometimes not. Made progress when recreated the controllers when recycling the views. But still not fixed. Followed the recommended way to add/remove view controllers https://guides.codepath.com/ios/Adding-and-Removing-Child-View-Controllers but it doesn't work. Seems to work for the UINavigationController but not the UITabBarController. Tried so many different combinations. Arrived at this one that works for most scenarios. But think there must be a problem with React Native - especially as controllers are empty when unmount called in TabBar (when doing the live reload with useEffect)