Survived live/reloads with mixed segmented tab
Had a NavigationStack in one segmented tab and non-NavigationStack in another segmented tab. Start on the stack tab and reload a couple of times (typing R) and would get a black screen. Tried this TabBarItem change a while back and it did make a difference in some scenarios but didn't think needed it after the latest recycle implementation. But it fixes this scenario so adding back