Survived live/reload when state navigator changes
In native twitter when edit the Tabs.js (change a tab title) and then navigate it threw an error. The live reload reran the usememo's and so the state navigators were created again. This temporarily reset the keys to [] and the native side threw an error when trying to navigate back to 0!!
Changed to not update keys if latest state navigator has empty context. This scenario only happens with live/reload - it still isn't quite right because it will navigate back to the first state and ignore the click - only first time, repeating the process seems to change the timings and the click works.
Generally it's better to define the state navigator outside the component so that state context is preserved across reloads - even real reloads like changing the phone's language.
Changed on NavigationMotion for consistency although didn't test live/reload