Set size when search view bounds change
In the old architecture called setSize on the bridge but this isn't there anymore!? Asked a question about the migration path on the working group https://github.com/reactwg/react-native-new-architecture/discussions/36 but no answer. So manually applied the width and height in the style. Not sure if this is equivalent (don't think it is because sometimes see a glitch where width isn't right?!) but can't see another way. Checked what React Native modal does because that's where copied setSize from originally - but this uses native state and that isn't possible without manually writing the generated files https://github.com/reactwg/react-native-new-architecture/discussions/31#discussioncomment-2716244 so not going down that road