Implemented placeholder prop on search bar
If placeholder not set then mustn't set it otherwise it wipes out the default value. But then want to set it if it was set and now isn't otherwise doesn't remove old value. This stuff is too hard to do with old and new props. When recycling the props haven't changed but still need to apply the value.
Anyway, decided to give it a default value from js. Fixes current bug of adding and removing placeholder not going back to default - setting to nil doesn't go back to default