Added title bar to navigation item title view
Got exception when navigating back from detail in the zoom example. The NavigationBar, in its unmount, tries to remove the title view from its subviews but it's not there anymore. It got moved when it was added to the titleview! In the old architecture this didn't happen because it was added and removed from react subviews instead (and react subviews moved to subviews). So mimicked that behaviour by removing child from superview (instead of removing from subviews)