Implemented anchor prop on ext/fab component
The new arch doesn't support findNodeHandle - so can't get the native id of a component in React. So changed anchor to accept type of component to anchor to. e.g., 'navigationBar'. This works because FAB and anchored component must be siblings under CoordinatorLayout parent so don't have to look far to find the anchor