Reverted forcing left position of title bar
It doesn't work reliably. The problem happens on orientation change with the title bar going back to left 0. Without this left change the problem always happens - prefer that for now. Should raise a react native bug because this is a change from old architecture. Might be because new arch doesn't call needsCustomLayoutForChildren?!