Took font/badgeColor props from any tab bar item
On  iOS >= 13 the font and badgeColor props are not specific to a tab bar item. Was taking them from only the first tab. So could be confusion if have badgeColor prop on second tab bar item and has no effect. So reduce over tab bar items and took the latest non-empty value for each prop