Ensured collapsing tabs don't block nav bar (RTL)
The RTL layout tabs are different so couldn't move the layout fix over to these. So changed both so they match. On attached to window set invisible if not already attached to view pager. Then set to visible if attach to pager. This works if top or bottom tabs (should do but haven't checked)