Ensured stack renders latest scene and props
Destructuring allScenes from current created a stale closure. So the renderScene didn't have the latests scenes and props. Changing it to allScenes.current throughout did work. But there was still a problem that allScenes updated in useEffect so it was always behind. Need allScenes to be updated as part of the render so it reflects the latest scenes in children