Pointed at codegen because didn't work in ios
Copying over the codegen didn't work for ios. Ended up with redefinition of EventEmitters errors https://github.com/reactwg/react-native-new-architecture/discussions/71#discussioncomment-3672029. There's no custom build on ios so it includes the copied over codegen as well as the actual codegen so it gets included twice. Can't just exclude it on ios because the includes on shadownodes point to local "EventEmitters.h'.
So went with the react native screens approach of pointing at the codegen instead of copying it over https://github.com/reactwg/react-native-new-architecture/discussions/71#discussioncomment-3656196. It works on android and should work on ios (haven't tested yet) because it's just the hand-written files now