Got it working on android and ios!!
Still copying code from react native screens example https://github.com/reactwg/react-native-new-architecture/discussions/71#discussioncomment-3656196
Couldn't work out how to get the ComponentDescriptor in the generated rncli.cpp because it only includes ComponentDescriptors.h. Noticed it also includes navigationreactnative.h - so the trick is to include the hand-written descriptors in that file!! React native screens did this just didn't notice or realise what it was for. It's very sneaky!
This allows it to work on android and ios!! And with lower maintenance because not copying over hand-generated - apart from navigationreactnative.h and .cpp. Had to exclude these files from the podspec otherwise doesn't build on ios