Added back image to navigation bar
This is tricky because the back image has to be set in viewWillAppear otherwise the default image shows during the transition. But react native loads the image asynchronously so have to delay the transition until the image has loaded.
Added a block function so the navigation bar can tell the stack when the image did load. Added a 100s bailout so the stack doesn't wait forever to start the transition. The image is loaded from the file system so should be enough time