Moved navigation completion outside of loop
Not sure why it was inside the loop - guess it made no difference. But now, don't want to add didLoad blocks to all controllers when fluently navigating. It's enough to add to the last controller (need to check though) because then fluently navigated back images will have plenty of time to load