Prevented repeated firing of completeNavigation
On the new arch (and maybe old?) would get the error saying trying to add the same view controller more than once. That was because the completeNavigation was firing twice. Not 100% on why - but the image loads faster on new arch, usually before the navigation happens - so it was going into the 0.1 NSEC callback part. Then it would sometimes later go into the imageDidLoad and call it a second time!?!
Not sure if it's recycling related. But tidied up the mechanism that prevents repeated calls - use a __block scope bool instead. Also gave up clearing the didLoad block. It's ok to call again because the complete check prevents damage.
Also changed the indicator from imageOn to imageLoading so that can clear it when the image loads - and so the new arch can immediately complete the navigation